Refactored bootcamp
