+++ public/locales/bg/common.json
@@ -17,10 +17,12 @@
     },
     "bootcamp": {
       "index": "Камп",
-      "create": "Създай държава",
-      "list": "Всички държави",
       "students": "Всички студенти",
-      "animals": "Всички животни",
+      "pets": {
+        "index": "Домашни любимци",
+        "create": "Добави нов",
+        "all": "Всички"
+      },
       "dashboard": "Админ панел"
     },
     "admin": {
