--- src/components/bootcamp/Drawer.tsx
@@ -1,205 +1,0 @@
-import { useState, useEffect } from 'react'
-import { styled, useTheme, Theme, CSSObject } from '@mui/material/styles'
-import makeStyles from '@mui/styles/makeStyles'
-import Box from '@mui/material/Box'
-import MuiDrawer from '@mui/material/Drawer'
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from '@mui/material/AppBar'
-import Toolbar from '@mui/material/Toolbar'
-import List from '@mui/material/List'
-import CssBaseline from '@mui/material/CssBaseline'
-import Typography from '@mui/material/Typography'
-import Divider from '@mui/material/Divider'
-import IconButton from '@mui/material/IconButton'
-import MenuIcon from '@mui/icons-material/Menu'
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft'
-import ChevronRightIcon from '@mui/icons-material/ChevronRight'
-import ListItemIcon from '@mui/material/ListItemIcon'
-import ListItemText from '@mui/material/ListItemText'
-import PetsIcon from '@mui/icons-material/Pets'
-import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline'
-import { ListItemButton } from '@mui/material'
-import { useRouter } from 'next/router'
-
-const drawerWidth = 240
-
-const openedMixin = (theme: Theme): CSSObject => ({
-  width: drawerWidth,
-  transition: theme.transitions.create('width', {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.enteringScreen,
-  }),
-  overflowX: 'hidden',
-})
-
-const closedMixin = (theme: Theme): CSSObject => ({
-  transition: theme.transitions.create('width', {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  overflowX: 'hidden',
-  width: `calc(${theme.spacing(7)} + 1px)`,
-  [theme.breakpoints.up('sm')]: {
-    width: `calc(${theme.spacing(9)} + 1px)`,
-  },
-})
-
-const DrawerHeader = styled('div')(({ theme }: { theme: Theme }) => ({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'flex-end',
-  padding: theme.spacing(0, 1),
-  ...theme.mixins.toolbar,
-}))
-
-interface AppBarProps extends MuiAppBarProps {
-  open?: boolean
-}
-
-const AppBar = styled(MuiAppBar, {
-  shouldForwardProp: (prop: string) => prop !== 'open',
-})<AppBarProps>(({ theme, open }: { theme: Theme; open: boolean }) => ({
-  zIndex: 1,
-  color: '#fff',
-  transition: theme.transitions.create(['width', 'margin'], {
-    easing: theme.transitions.easing.sharp,
-    duration: theme.transitions.duration.leavingScreen,
-  }),
-  ...(open && {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  }),
-}))
-
-const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop: string) => prop !== 'open' })(
-  ({ theme, open }: { theme: Theme; open: boolean }) => ({
-    zIndex: 0,
-    width: drawerWidth,
-    flexShrink: 0,
-    whiteSpace: 'nowrap',
-    boxSizing: 'border-box',
-    ...(open && {
-      ...openedMixin(theme),
-      '& .MuiDrawer-paper': {
-        position: 'absolute',
-        width: drawerWidth,
-        transition: theme.transitions.create('width', {
-          easing: theme.transitions.easing.sharp,
-          duration: theme.transitions.duration.enteringScreen,
-        }),
-        overflowX: 'hidden',
-      },
-    }),
-    ...(!open && {
-      ...closedMixin(theme),
-      '& .MuiDrawer-paper': {
-        position: 'absolute',
-        transition: theme.transitions.create('width', {
-          easing: theme.transitions.easing.sharp,
-          duration: theme.transitions.duration.leavingScreen,
-        }),
-        overflowX: 'hidden',
-        width: `calc(${theme.spacing(7)} + 1px)`,
-        [theme.breakpoints.up('sm')]: {
-          width: `calc(${theme.spacing(9)} + 1px)`,
-        },
-      },
-    }),
-  }),
-)
-
-const useStyles = makeStyles(() => {
-  return {
-    listItem: {
-      paddingLeft: '22px',
-    },
-  }
-})
-
-type Props = {
-  children: React.ReactNode
-}
-
-export default function MiniDrawer({ children }: Props) {
-  const theme = useTheme()
-  const router = useRouter()
-  const [open, setOpen] = useState(false)
-  const [selectIndex, setSelectIndex] = useState(1)
-  const classes = useStyles()
-
-  useEffect(() => {
-    const url = router.asPath
-    if (url.endsWith('create')) {
-      setSelectIndex(2)
-    }
-  }, [])
-
-  const handleDrawerOpen = () => {
-    setOpen(true)
-  }
-
-  const handleDrawerClose = () => {
-    setOpen(false)
-  }
-
-  return (
-    <Box sx={{ display: 'flex', position: 'absolute', width: '100%' }}>
-      <CssBaseline />
-      <AppBar position="absolute" open={open}>
-        <Toolbar>
-          <IconButton
-            color="inherit"
-            aria-label="open drawer"
-            onClick={handleDrawerOpen}
-            edge="start"
-            sx={{
-              marginRight: '36px',
-              ...(open && { display: 'none' }),
-            }}>
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="h6" noWrap component="div">
-            Pets dashboard
-          </Typography>
-        </Toolbar>
-      </AppBar>
-      <Drawer variant="permanent" open={open}>
-        <DrawerHeader>
-          <IconButton onClick={handleDrawerClose}>
-            {theme.direction === 'rtl' ? <ChevronRightIcon /> : <ChevronLeftIcon />}
-          </IconButton>
-        </DrawerHeader>
-        <Divider />
-        <List>
-          <ListItemButton
-            component="a"
-            href="/bootcamp/animals"
-            className={classes.listItem}
-            selected={selectIndex == 1}>
-            <ListItemIcon>
-              <PetsIcon />
-            </ListItemIcon>
-            <ListItemText primary="All" />
-          </ListItemButton>
-          <ListItemButton
-            component="a"
-            href="/bootcamp/animals/create"
-            className={classes.listItem}
-            selected={selectIndex == 2}>
-            <ListItemIcon>
-              <AddCircleOutlineIcon />
-            </ListItemIcon>
-            <ListItemText primary="Create" />
-          </ListItemButton>
-        </List>
-      </Drawer>
-      <Box component="main" sx={{ flexGrow: 1, p: 3, minHeight: '400px' }}>
-        <DrawerHeader />
-        {children}
-      </Box>
-    </Box>
-  )
-}
