+++ src/components/layout/nav/DashboardMenu.tsx
@@ -1,71 +1,69 @@
-import DashboardIcon from '@mui/icons-material/Dashboard'
-import { List, ListItemButton, ListItemIcon, ListItemText } from '@mui/material'
-import BasicPopover from 'components/bootcamp/BasicPopover'
+import { List } from '@mui/material'
 import PetsIcon from '@mui/icons-material/Pets'
 import AddBoxIcon from '@mui/icons-material/AddBox'
-import RestaurantIcon from '@mui/icons-material/Restaurant'
-import FastfoodIcon from '@mui/icons-material/Fastfood'
-import CakeIcon from '@mui/icons-material/Cake'
-import IcecreamIcon from '@mui/icons-material/Icecream'
-import LunchDiningIcon from '@mui/icons-material/LunchDining'
+import DashboardIcon from '@mui/icons-material/Dashboard'
+import { ExpandLess, ExpandMore } from '@mui/icons-material'
+import Link from 'next/link'
+import { ListItemButton, ListItemIcon, ListItemText, Collapse } from '@mui/material'
+import { DrawerStore } from 'stores/DrawerStore'
+import { observer } from 'mobx-react'
+import { useRouter } from 'next/router'
+import { routes } from 'common/routes'
+import { useTranslation } from 'next-i18next'
 
-export default function DashboardMenu() {
+export default observer(function DashboardMenu() {
+  const { t } = useTranslation()
+  const { subMenuIsOpen, toggleSubMenu } = DrawerStore
+  const router = useRouter()
+
   return (
-    <List>
-      <ListItemButton component="a" href="/bootcamp/dashboard">
+    <List
+      sx={{ width: '100%', maxWidth: 360, bgcolor: 'background.paper' }}
+      component="nav"
+      aria-labelledby="nested-list-subheader">
+      <Link passHref href={routes.bootcamp.dashboard.index}>
+        <ListItemButton component="a" selected={router.asPath == routes.bootcamp.dashboard.index}>
+          <ListItemIcon>
+            <DashboardIcon />
+          </ListItemIcon>
+          <ListItemText primary={t('nav.bootcamp.students')} />
+        </ListItemButton>
+      </Link>
+      <ListItemButton
+        onClick={toggleSubMenu}
+        selected={router.asPath.startsWith(routes.bootcamp.dashboard.pets) && !subMenuIsOpen}>
         <ListItemIcon>
-          <DashboardIcon />
+          <PetsIcon />
         </ListItemIcon>
-        <ListItemText primary="Dashboard" />
+        <ListItemText primary={t('nav.bootcamp.pets.index')} />
+        {subMenuIsOpen ? <ExpandLess /> : <ExpandMore />}
       </ListItemButton>
-      <BasicPopover title="Pets" icon={<PetsIcon />}>
-        <List>
-          <ListItemButton component="a" href="/bootcamp/dashboard/pets">
-            <ListItemIcon>
-              <PetsIcon />
-            </ListItemIcon>
-            <ListItemText primary="All pets" />
-          </ListItemButton>
-          <ListItemButton component="a" href="/bootcamp/dashboard/pets/create">
-            <ListItemIcon>
-              <AddBoxIcon />
-            </ListItemIcon>
-            <ListItemText primary="Create pet" />
-          </ListItemButton>
+      <Collapse in={subMenuIsOpen} timeout="auto" unmountOnExit>
+        <List component="div" disablePadding>
+          <Link passHref href={routes.bootcamp.dashboard.pets}>
+            <ListItemButton
+              sx={{ pl: 4 }}
+              component="a"
+              selected={router.asPath == routes.bootcamp.dashboard.pets}>
+              <ListItemIcon>
+                <PetsIcon />
+              </ListItemIcon>
+              <ListItemText primary={t('nav.bootcamp.pets.all')} />
+            </ListItemButton>
+          </Link>
+          <Link passHref href={routes.bootcamp.dashboard.createPet}>
+            <ListItemButton
+              sx={{ pl: 4 }}
+              component="a"
+              selected={router.asPath == routes.bootcamp.dashboard.createPet}>
+              <ListItemIcon>
+                <AddBoxIcon />
+              </ListItemIcon>
+              <ListItemText primary={t('nav.bootcamp.pets.create')} />
+            </ListItemButton>
+          </Link>
         </List>
-      </BasicPopover>
-      <BasicPopover title="Food" icon={<RestaurantIcon />}>
-        <List>
-          <ListItemButton component="a" href="/bootcamp/dashboard">
-            <ListItemIcon>
-              <RestaurantIcon />
-            </ListItemIcon>
-            <ListItemText primary="Restaurants" />
-          </ListItemButton>
-          <BasicPopover title="Fast food" icon={<FastfoodIcon />}>
-            <List>
-              <ListItemButton component="a" href="/bootcamp/dashboard">
-                <ListItemIcon>
-                  <CakeIcon />
-                </ListItemIcon>
-                <ListItemText primary="Cakes" />
-              </ListItemButton>
-              <ListItemButton component="a" href="/bootcamp/dashboard">
-                <ListItemIcon>
-                  <IcecreamIcon />
-                </ListItemIcon>
-                <ListItemText primary="Ice cream" />
-              </ListItemButton>
-              <ListItemButton component="a" href="/bootcamp/dashboard">
-                <ListItemIcon>
-                  <LunchDiningIcon />
-                </ListItemIcon>
-                <ListItemText primary="Burgers" />
-              </ListItemButton>
-            </List>
-          </BasicPopover>
-        </List>
-      </BasicPopover>
+      </Collapse>
     </List>
   )
-}
+})
