--- src/pages/bootcamp/animals/index.tsx
@@ -1,13 +1,0 @@
-import AnimalsPage from 'components/bootcamp/AnimalsPage'
-import { GetServerSideProps } from 'next'
-import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
-
-export const getServerSideProps: GetServerSideProps = async ({ locale }) => {
-  return {
-    props: {
-      ...(await serverSideTranslations(locale ?? 'bg', ['common'])),
-    },
-  }
-}
-
-export default AnimalsPage
