+++ src/pages/bootcamp/dashboard/pets/create.tsx
@@ -1,4 +1,4 @@
-import CreateAnimalForm from 'components/bootcamp/CreateAnimalForm'
+import CreatePetForm from 'components/bootcamp/CreatePetForm'
 import { GetStaticProps } from 'next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { Container } from '@mui/material'
@@ -8,7 +8,7 @@ export default function CreatePet() {
   return (
     <DashboardLayout title="Create pet">
       <Container>
-        <CreateAnimalForm redirectUrl="/bootcamp/dashboard/pets" />
+        <CreatePetForm redirectUrl="/bootcamp/dashboard/pets" />
       </Container>
     </DashboardLayout>
   )
